requirements.txt
# Core LLM / LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-nvidia-ai-endpoints>=0.3.0

# Serving
langserve[all]>=0.3.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sse-starlette>=2.0.0

# Frontend
gradio>=4.0.0

# Vector Store
faiss-cpu>=1.7.4

# Document Processing
pymupdf>=1.24.0
python-docx>=1.1.0

# Data / ML Utilities
numpy>=1.26.0
scikit-learn>=1.4.0
pydantic>=2.6.0

# Environment / Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.9
rich>=13.7.0
httpx>=0.27.0
